Moved to single function

+++ src/NzbDrone.Mono/Disk/DiskProvider.cs
@@ -171,7 +171,7 @@ namespace NzbDrone.Mono.Disk
             }
             else if (PlatformInfo.GetVersion() > new Version(6, 0) && (!FileExists(destination) || overwrite))
             {
-                CopyFileChecked(source, destination, overwrite);
+                TransferFilePatched(source, destination, overwrite, false);
             }
             else
             {
@@ -179,111 +179,6 @@ namespace NzbDrone.Mono.Disk
             }
         }
 
-        private void CopyFileChecked(string source, string destination, bool overwrite)
-        {
-            // Mono 6.x throws errors if permissions or timestamps cannot be set
-            // - In 6.0 it'll leave a full length file
-            // - In 6.6 it'll leave a zero length file
-            // Catch the exception and attempt to handle these edgecases
-
-            // Our check depends on the destination not existing
-            if (FileExists(destination) && overwrite)
-            {
-                DeleteFile(destination);
-            }
-
-            try
-            {
-                base.CopyFileInternal(source, destination, overwrite);
-            }
-            catch (UnauthorizedAccessException)
-            {
-                var srcInfo = new FileInfo(source);
-                var dstInfo = new FileInfo(destination);
-                var exists = dstInfo.Exists && srcInfo.Exists;
-                if (exists && dstInfo.Length == 0 && srcInfo.Length != 0)
-                {
-                    // mono >=6.6 bug: zero length file since chmod happens at the start
-                    _logger.Debug("Mono failed to copy file likely due to known mono bug, attempting to copy directly '{0}' -> '{1}'", source, destination);
-
-                    try
-                    {
-                        _logger.Trace("Copying content from {0} to {1} ({2} bytes)", source, destination, srcInfo.Length);
-                        using (var srcStream = new FileStream(source, FileMode.Open, FileAccess.Read))
-                        using (var dstStream = new FileStream(destination, FileMode.Create, FileAccess.Write))
-                        {
-                            srcStream.CopyTo(dstStream);
-                        }
-                    }
-                    catch
-                    {
-                        // If it fails again then bail
-                        throw;
-                    }
-
-                    try
-                    {
-                        dstInfo.LastWriteTimeUtc = srcInfo.LastWriteTimeUtc;
-                    }
-                    catch
-                    {
-                        _logger.Debug("Unable to change last modified date for {0}, skipping.", destination);
-                    }
-                }
-                else if (exists && dstInfo.Length == srcInfo.Length)
-                {
-                    // mono 6.0, 6.4 bug: full length file since utime and chmod happens at the end
-                    _logger.Debug("Mono failed to copy file likely due to known mono bug, attempting to copy directly '{0}' -> '{1}'", source, destination);
-
-                    // Don't take risks, UnauthorizedAccess can happen due to various reasons. Check the file.
-                    var checkLength = (int)Math.Min(64 * 1204, dstInfo.Length);
-                    if (checkLength > 0)
-                    {
-                        var srcData = new byte[checkLength];
-                        var dstData = new byte[checkLength];
-
-                        _logger.Trace("Checking last {0} bytes in {1}", checkLength, destination);
-
-                        using (var srcStream = new FileStream(source, FileMode.Open, FileAccess.Read))
-                        using (var dstStream = new FileStream(destination, FileMode.Open, FileAccess.Read))
-                        {
-                            srcStream.Position = srcInfo.Length - checkLength;
-                            dstStream.Position = dstInfo.Length - checkLength;
-
-                            srcStream.Read(srcData, 0, checkLength);
-                            dstStream.Read(dstData, 0, checkLength);
-                        }
-
-                        for (var i = 0; i < checkLength; i++)
-                        {
-                            if (srcData[i] != dstData[i])
-                            {
-                                // Files aren't the same, the UnauthorizedAccess was unrelated
-                                _logger.Trace("Copy was incomplete, rethrowing original error");
-                                throw;
-                            }
-                        }
-
-                        _logger.Trace("Copy was complete, finishing copy ourselves");
-                    }
-
-                    try
-                    {
-                        dstInfo.LastWriteTimeUtc = srcInfo.LastWriteTimeUtc;
-                    }
-                    catch
-                    {
-                        _logger.Debug("Unable to change last modified date for {0}, skipping.", destination);
-                    }
-                }
-                else
-                {
-                    // Unrecognized situation, the UnauthorizedAccess was unrelated
-                    throw;
-                }
-            }
-        }
-        
         protected override void MoveFileInternal(string source, string destination)
         {
             var sourceInfo = UnixFileSystemInfo.GetFileSystemEntry(source);
@@ -321,21 +216,21 @@ namespace NzbDrone.Mono.Disk
             }
             else if (PlatformInfo.GetVersion() > new Version(6, 0) && !FileExists(destination))
             {
-                MoveFileChecked(source, destination);
+                TransferFilePatched(source, destination, false, true);
             }
             else
             {
                 base.MoveFileInternal(source, destination);
             }
         }
-
-        private void MoveFileChecked(string source, string destination)
+        
+        private void TransferFilePatched(string source, string destination, bool overwrite, bool move)
         {
             // Mono 6.x throws errors if permissions or timestamps cannot be set
             // - In 6.0 it'll leave a full length file
             // - In 6.6 it'll leave a zero length file
             // Catch the exception and attempt to handle these edgecases
-            
+
             try
             {
                 base.MoveFileInternal(source, destination);
@@ -348,7 +243,7 @@ namespace NzbDrone.Mono.Disk
                 if (exists && dstInfo.Length == 0 && srcInfo.Length != 0)
                 {
                     // mono >=6.6 bug: zero length file since chmod happens at the start
-                    _logger.Debug("Mono failed to move file likely due to known mono bug, attempting to copy&move directly. '{0}' -> '{1}'", source, destination);
+                    _logger.Debug("Mono failed to {3} file likely due to known mono bug, attempting to {3} directly. '{0}' -> '{1}'", source, destination, move ? "move" : "copy");
 
                     try
                     {
@@ -374,12 +269,16 @@ namespace NzbDrone.Mono.Disk
                         _logger.Debug("Unable to change last modified date for {0}, skipping.", destination);
                     }
 
-                    File.Delete(source);
+                    if (move)
+                    {
+                        _logger.Trace("Removing source file {0}", source);
+                        File.Delete(source);
+                    }
                 }
                 else if (exists && dstInfo.Length == srcInfo.Length)
                 {
                     // mono 6.0, 6.4 bug: full length file since utime and chmod happens at the end
-                    _logger.Debug("Mono failed to move file likely due to known mono bug, attempting to copy&move directly. '{0}' -> '{1}'", source, destination);
+                    _logger.Debug("Mono failed to {3} file likely due to known mono bug, attempting to {3} directly. '{0}' -> '{1}'", source, destination, move ? "move" : "copy");
 
                     // Check at least part of the file since UnauthorizedAccess can happen due to legitimate reasons too
                     var checkLength = (int)Math.Min(64 * 1204, dstInfo.Length);
@@ -410,7 +309,7 @@ namespace NzbDrone.Mono.Disk
                             }
                         }
 
-                        _logger.Trace("Copy was complete, finishing move operation");
+                        _logger.Trace("Copy was complete, finishing {0} operation", move ? "move" : "copy");
                     }
 
                     try
@@ -422,7 +321,11 @@ namespace NzbDrone.Mono.Disk
                         _logger.Debug("Unable to change last modified date for {0}, skipping.", destination);
                     }
 
-                    File.Delete(source);
+                    if (move)
+                    {
+                        _logger.Trace("Removing source file {0}", source);
+                        File.Delete(source);
+                    }
                 }
                 else
                 {
